Rewrote all of the windows socket and threading functions to use our new error handling

+++ netlib/headers/netemu_socket.h
@@ -233,7 +233,7 @@ typedef unsigned int NETEMU_SOCKET;
 	unsigned long netemu_ntohl(unsigned long value);
 
 	/* Converts an unsigned long from network order to host order. */
-	unsigned long netemu_htons(unsigned long value);
+	unsigned short netemu_htons(unsigned short value);
 
 #ifdef	__cplusplus
 }
